+++ quilt-sync/src-tauri/src/model.rs
@@ -193,7 +193,7 @@ pub trait QuiltModel {
         metadata: Option<serde_json::Value>,
         workflow: Option<quilt::manifest::Workflow>,
         host_config: Option<HostConfig>,
-    ) -> Result<quilt::PushOutcome, Error> {
+    ) -> Result<quilt::PublishOutcome, Error> {
         Ok(package
             .publish(message, metadata, workflow, host_config)
             .await?)
@@ -545,7 +545,7 @@ pub async fn package_publish(
     metadata: &str,
     workflow: Option<String>,
     host_config: Option<HostConfig>,
-) -> Result<quilt::PushOutcome, Error> {
+) -> Result<quilt::PublishOutcome, Error> {
     debug!(
         "Publishing the package.\nNamespace: {},\nmessage: {},\nuser_meta: {},\nworkflow: {:?}",
         namespace, message, metadata, workflow